requirements.txt
# Core dependencies
fastapi
uvicorn[standard]
pydantic

# Vector database
weaviate-client

# Document parsing
python-dotenv
PyPDF2
pdf2image
Pillow

# M3-Optimized ML Stack (Apple Silicon)
torch==2.5.1  # Pinned! 2.6.0 has MPS bugs
transformers>=4.45.0
colpali-engine>=0.3.0
qwen-vl-utils

# Apple Silicon Optimizations
mlx>=0.18.0  # For future native MLX support
accelerate  # Automatic device mapping

# API utilities
python-multipart
requests

# Dual-Model Analysis Sub-Agent
openai>=1.50.0  # GPT-5.1 Responses API (high reasoning)
google-genai>=1.50.0  # Gemini 3 Pro (1M context window)

# DSPy for prompt optimization
dspy-ai

# Experiment tracking (optional)
mlflow
